isAdmin works on frontend

+++ seekr/src/components/item/Item.jsx
@@ -34,13 +34,14 @@ export class Item extends Component {
     const { id, name, found, desc, location, images, timestamp, tags, username, email} = this.props.item;
     var lat = location.coordinates[0]
     var lng = location.coordinates[1]
+
     var url = "https://www.google.com/maps/place/" + lat.toString(10) + "+" + lng.toString(10)
     return (
       <div onClick={() => this.modalHandler(true)} style={{}}>
         <Card border={found ? 'success' : 'warning'} style={{ textAlign: 'left', width: "16rem", margin: '1em', borderWidth: '0.25em'}}>
           <Card.Title style={{ margin: '1em 0.8em 0.5em'}}>{ name }</Card.Title>
           <Tags tags={ItemTags.getStrings(tags).split(',')} ></Tags> 
-          {email === this.props.cur_email ?  <Button style={{ margin: '1em 0.5em 1em 1em'}} variant="danger" onClick={this.props.deleteItem.bind(this, id)}>Delete</Button> : null}
+          {((email === this.props.cur_email) || this.props.isAdmin === 'true') ?  <Button style={{ margin: '1em 0.5em 1em 1em'}} variant="danger" onClick={this.props.deleteItem.bind(this, id)}>Delete</Button> : null}
           { images.length !== 0 ? this.getImage(images) : <span></span>}
           <Card.Footer>
             <small className="text-muted">{ found ? "Found item, " : "Lost item, " }</small>
