+++ seekr/src/components/item/Items.jsx
@@ -3,10 +3,9 @@ import Item from './Item';
 import { AuthContext } from "../helper/AuthContext";
 
 const Items = ({items, deleteItem}) => {
-  const { email } = useContext(AuthContext);
-
+  const { email, isAdmin } = useContext(AuthContext);
   return items.map((item) => (
-    <Item key={item.id} item={item} deleteItem={deleteItem} cur_email={email} />
+    <Item key={item.id} item={item} deleteItem={deleteItem} cur_email={email} isAdmin={isAdmin} />
   ));
 };
